# src/app/api/filaments/[id]/route.ts
import { NextRequest, NextResponse } from "next/server";
import dbConnect from "@/lib/mongodb";
import Filament, { IFilament } from "@/models/Filament";
import "@/models/Nozzle";
import "@/models/Printer";
import { resolveFilament, hasVariants } from "@/lib/resolveFilament";
import { getErrorMessage, errorResponse } from "@/lib/apiErrorHandler";

export async function GET(
  _request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  try {
    await dbConnect();
    const { id } = await params;
    const filament = await Filament.findOne({ _id: id, _deletedAt: null })
      .populate("compatibleNozzles")
      .populate("calibrations.nozzle")
      .populate("calibrations.printer")
      .lean();
    if (!filament) {
      return errorResponse("Not found", 404);
    }

    // If this is a variant, resolve inherited values from parent
    let resolved: IFilament | ReturnType<typeof resolveFilament> = filament;
    if (filament.parentId) {
      const parent = await Filament.findOne({ _id: filament.parentId, _deletedAt: null })
        .populate("compatibleNozzles")
        .populate("calibrations.nozzle")
        .populate("calibrations.printer")
        .lean();
      resolved = resolveFilament(filament, parent);
    }

    // If this is a parent, include its variants
    const variants = await Filament.find({ parentId: id, _deletedAt: null })
      .select("name color cost")
      .sort({ name: 1 })
      .lean();

    return NextResponse.json({ ...resolved, _variants: variants });
  } catch (err) {
    return errorResponse("Failed to fetch filament", 500, getErrorMessage(err));
  }
}

export async function PUT(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  let body: any;
  try {
    body = await request.json();
  } catch {
    return errorResponse("Invalid JSON in request body", 400);
  }

  try {
    await dbConnect();
    const { id } = await params;
    delete body._id;
    delete body._deletedAt;
    delete body.createdAt;
    delete body.updatedAt;
    delete body.__v;
    delete body.instanceId;
    delete body.syncId;

    // Validate parentId if provided
    if (body.parentId) {
      const parent = await Filament.findOne({ _id: body.parentId, _deletedAt: null }).lean();
      if (!parent) {
        return errorResponse("Parent filament not found", 400);
      }
      // Prevent circular references
      if (parent.parentId) {
        return errorResponse("Cannot set a variant as parent (no nested inheritance)", 400);
      }
      // Prevent self-reference
      if (body.parentId === id) {
        return errorResponse("Cannot be your own parent", 400);
      }
      // Prevent converting a parent to a variant while it has children
      const variantCount = await Filament.countDocuments({ parentId: id, _deletedAt: null });
      if (variantCount > 0) {
        return errorResponse("Cannot set parent on a filament that has variants — remove variants first", 400);
      }
    }

    const filament = await Filament.findOneAndUpdate(
      { _id: id, _deletedAt: null },
      body,
      { new: true, runValidators: true }
    ).lean();
    if (!filament) {
      return errorResponse("Not found", 404);
    }
    return NextResponse.json(filament);
  } catch (err) {
    return errorResponse("Failed to update filament", 500, getErrorMessage(err));
  }
}

/**
 * POST /api/filaments/:nameOrId
 *
 * Sync a filament preset back from PrusaSlicer. The param can be a
 * URL-encoded preset name (e.g. "The%20K8%20PC") or a MongoDB ObjectId.
 *
 * Body: { name: string, config: Record<string, string> }
 *
 * Finds the filament by name (falling back to _id), then merges the
 * incoming config keys into the filament's `settings` bag.
 */
export async function POST(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  let body: any;
  try {
    body = await request.json();
  } catch {
    return errorResponse("Invalid JSON in request body", 400);
  }

  try {
    await dbConnect();
    const { id } = await params;
    const config: Record<string, string> = body.config || {};

    if (!config || Object.keys(config).length === 0) {
      return errorResponse("No config provided", 400);
    }

    // Try to find by name first (PrusaSlicer sends URL-encoded name),
    // then fall back to ObjectId
    const decodedName = decodeURIComponent(id);
    let filament = await Filament.findOne({ name: decodedName, _deletedAt: null });
    if (!filament && /^[a-f0-9]{24}$/i.test(id)) {
      filament = await Filament.findOne({ _id: id, _deletedAt: null });
    }

    if (!filament) {
      return errorResponse(`Filament not found: ${decodedName}`, 404);
    }

    // Reverse-map PrusaSlicer INI keys → structured DB fields
    const update: Record<string, unknown> = {};
    const temps: Record<string, unknown> = {};

    // Core fields
    if (config.filament_type) update.type = config.filament_type;
    if (config.filament_vendor) update.vendor = config.filament_vendor;
    if (config.filament_colour) update.color = config.filament_colour;
    if (config.filament_diameter) { const v = parseFloat(config.filament_diameter); if (!isNaN(v)) update.diameter = v; }
    if (config.filament_density) { const v = parseFloat(config.filament_density); if (!isNaN(v)) update.density = v; }
    if (config.filament_cost) { const v = parseFloat(config.filament_cost); if (!isNaN(v)) update.cost = v; }
    if (config.filament_spool_weight) { const v = parseFloat(config.filament_spool_weight); if (!isNaN(v)) update.spoolWeight = v; }
    if (config.filament_max_volumetric_speed) { const v = parseFloat(config.filament_max_volumetric_speed); if (!isNaN(v)) update.maxVolumetricSpeed = v; }

    // Temperatures
    if (config.temperature) { const v = parseInt(config.temperature); if (!isNaN(v)) temps.nozzle = v; }
    if (config.first_layer_temperature) { const v = parseInt(config.first_layer_temperature); if (!isNaN(v)) temps.nozzleFirstLayer = v; }
    if (config.bed_temperature) { const v = parseInt(config.bed_temperature); if (!isNaN(v)) temps.bed = v; }
    if (config.first_layer_bed_temperature) { const v = parseInt(config.first_layer_bed_temperature); if (!isNaN(v)) temps.bedFirstLayer = v; }

    // Shrinkage
    if (config.filament_shrinkage_compensation_xy) { const v = parseFloat(config.filament_shrinkage_compensation_xy); if (!isNaN(v)) update.shrinkageXY = v; }
    if (config.filament_shrinkage_compensation_z) { const v = parseFloat(config.filament_shrinkage_compensation_z); if (!isNaN(v)) update.shrinkageZ = v; }

    // Flags
    if (config.filament_soluble) update.soluble = config.filament_soluble === "1";
    if (config.filament_abrasive) update.abrasive = config.filament_abrasive === "1";

    // Merge temperatures into existing
    if (Object.keys(temps).length > 0) {
      const existing = (filament.temperatures as Record<string, unknown>) || {};
      update.temperatures = { ...existing, ...temps };
    }

    // Everything else goes into the settings bag
    const STRUCTURED_KEYS = new Set([
      "filament_type", "filament_vendor", "filament_colour", "filament_diameter",
      "filament_density", "filament_cost", "filament_spool_weight",
      "filament_max_volumetric_speed", "temperature", "first_layer_temperature",
      "bed_temperature", "first_layer_bed_temperature",
      "filament_shrinkage_compensation_xy", "filament_shrinkage_compensation_z",
      "filament_soluble", "filament_abrasive", "filament_settings_id",
    ]);
    const settings = (filament.settings as Record<string, unknown>) || {};
    for (const [key, value] of Object.entries(config)) {
      if (!STRUCTURED_KEYS.has(key)) {
        settings[key] = value;
      }
    }
    update.settings = settings;

    await Filament.findByIdAndUpdate(filament._id, { $set: update });

    return NextResponse.json({
      message: `Synced ${Object.keys(config).length} settings for "${decodedName}"`,
      filamentId: filament._id,
    });
  } catch (err) {
    return errorResponse("Failed to sync filament", 500, getErrorMessage(err));
  }
}

export async function DELETE(
  _request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  try {
    await dbConnect();
    const { id } = await params;

    // Prevent deleting a parent that has variants
    if (await hasVariants(Filament, id)) {
      return errorResponse(
        "Cannot delete a filament that has color variants. Delete the variants first.",
        400,
      );
    }

    const filament = await Filament.findOneAndUpdate(
      { _id: id, _deletedAt: null },
      { _deletedAt: new Date() },
      { new: true }
    ).lean();
    if (!filament) {
      return errorResponse("Not found", 404);
    }
    return NextResponse.json({ message: "Deleted" });
  } catch (err) {
    return errorResponse("Failed to delete filament", 500, getErrorMessage(err));
  }
}